Make it possible to compile OMC without backend
For bootstrapping and the bootstrapping tarball in particular, the
OMC backend is not needed. This patch adds stubs for (some) unused
parts of the compiler. Using the stubs requires setting an env. var
(done automatically when creating a new tarball).

Some packages were split up (SimCodeUtil, SimCodeFunctionUtil) in order
to avoid having references to equation systems in the compiler. Still
missing:
- CevalScript
- Interactive
- CodegenC
Some additional packages are in the wrong list at the moment.

An additional side-effect is that it should be possible to use new
language features in the backend parts by stubbing them out when
bootstrapping. The three phases will accomodate this:
- First phase from the tarball (OMC not needed).
- Second phase cannot use new language features, but does not compile
  the backend.
- Third phase now has an up-to-date OMC with stubbed backend, which is
  able to compile the backend with new language features.