Rewrote Absyn.pathString
The new implementation of pathString uses external object
System.StringAllocator to allocate a string as well as
System.stringAllocatorStringCopy to write into this external object
in a way that regular MetaModelica functions cannot. This avoids
a bunch of list creation and delimiting of string lists.

The size of the String is first detected by traversing the path,
which is needed in order to allocate a string of the correct length.
After this, the data is copied into the correct places.