Changes for bootstrapping header
OpenModelicaBootstrappingHeader.h is now treated as part of the tarball
for bootstrapping. This makes it possible to change Absyn.mo without
generating a new tarball.

The build system now generates special a-files for Compiler/runtime/
and Parser/ that uses the tarball version of the header for stage 1 of
the bootstrapping, and a fresh header for other objects. This file is
not copied to the build directory.