Remove FMI1 log-message
The reason for removing this log-message is that FMIL passes loggingOn
as true even if we call instantiate with loggingOn=false. The FMU would
work correctly when FMIL was not used, but the OM FMU import does use
FMIL and it is annoying to see the GUID when not debugging.