Experimental support for OPC UA embedded server
We support:

- OPC UA subscriptions for boolean/real variables
- Changing boolean/real variables (but not triggering an event)
- Boolean/Real aliases
- Not Real discrete variables and parameters
- Real-time synchronization support (via OPC or simflags: -rt 1.0)
- Having no stop-time (OPC only)
- Handling lowlatency kernel as root, allowing a tighter real-time
  frequency (best-case 100 Hz as regular user, 2 kHz as root without
  OPC communication or writing simulation results)

We no longer trigger time events at stopTime+1. This causes problems
when the stopTime is not used. Instead, we put a value of NAN for the
next sample event when there are no samples which effectively disables
time events.

There is a simple Python client for testing using a GUI. A testcase
for OPC is still missing.