Add a removeVerySimpleEquations pre-opt module
The reason for adding an additional early and faster remove simple
equations is that it speeds up the subsequent modules since they need
to traverse fewer equations and allocate less data. This gives upwards
of a 20% performance increase for medium-size models.

This is disabled by default and marked Experimental due to the used
algorithm not working for all cases.

It is only used for the preOptModules, but for systems depending on
LateInline, it could be possible to use it as the first post-opt module
as well.