Make error messages less verbose for performance
Do not create a prefixed path to string for every single variable that
is instantiated. Possibility: use a debugging flag to turn on more
diagnostics. This *should* be OK since we pass along SourceInfo to most
of the functions in the compiler, which is the main task of
Error.updateCurrentComponent. Also getting the full instance name
should be a minor detail.