Use templates to write to file directly
Used misc. templates in other templates instead of importing the string
conversion function in the templates. The output changes slightly since
we can no longer trim strings, etc.

Also introduced the possibility to use File and Tpl.Text objects in the
same package (used for JSON serialization, which is mostly directly to
file but also needs to printExpStr).