Cache typeviews in Susan
This improves performance a lot since the same typeview is added
multiple times and it does linear search or something, making the
performance of Susan scale n^2 or worse with the number of typeviews
included (and each imported template imports new typeviews).

Time to run CodegenFMUCppHpcom.tpl went from 8 minutes to a bit less
than a second, showing that scalability improves drastically.