Simplify of asub where the subscript is a range
This handles some obscure code that gets inlined. Ranges are not
simplified in the general case because they can be huge and usually are
used to iterate over some indexes without the need to construct an
array. Subscripts are different in that they are used to construct
arrays, so they should be safe to evaluate at compile-time.