Added optional Susan debugging help
There is a new flag -d=susanDebug which can generate a different
MetaModelica from the Susan code. It will give a stack-trace of the
point where a failure happens.

Use the debug-flag when compiling the Susan code to add extra try/catch
blocks everywhere (this slows down the code considerably and is not
intended for use in the production executable; only for debugging).