Fix for elabBuiltinSize of zero-size arrays
Previously, we had work-arounds for T_UNKNOWN being passed to size.
Since we fixed that, we now need to provide the correct number of
dimensions. This is either the expression type or the type from
elaboration; if we take the longer of the two we should get the
correct number of dimensions.