Handle inline where inputs are record fun-calls
Use DAE.RSUB() to handle general expressions. When doing inline of a
function call, we can use DAE.RSUB instead of failing; previously we
only handled component references passed to the call. Also added code
generation and simplifications for DAE.RSUB since it was only used
for MetaModelica previously.

This fixes some of the issues raised in ticket:4423.