Redesign SymbolTable handling
The SymbolTable has been moved out from the frontend completely. It is
now handled by a new module called SymbolTable, which uses a thread-
local variable to avoid passing around the symbol table. This makes the
handling of caching SCode much simpler.

Some functionality has been removed:
- Cached classes in the symbol table (we use the FCache instead)
- Instantiated functions (since they were not used before, and -d=gen
  is not used anyway)
- The LoadedFiles list (since the result was ignored anyway)
- OpenModelicaInternal.simplify (since the bootstrapping tests can run
  simplify instead)

Other functionality has been moved between frontend and backend files in
order to make the interface cleaner.

The handling of loadResource has been updated. The locations of
resources is now stored in the runtime in order to make the handling
of URIs easier to implement for FMI.

There are also fixes for e.g. getErrorString and runScript to not
generate error-messages during elaboration. This is done by making
getErrorString impure, and by making elabCall able to roll back messages
that were not at the top of the stack previously.

CevalFunction now stops evaluation on fail() like it does for
assert(false).

The frontend now fails when the function tree is not valid because there
seems to be some model(s) that manage to ignore failing to instantiate a
function.

We now handle assertions triggered in built-in functions that previously
printed to stderr by making them write to the Error.mo buffer instead.