Catch out of memory errors
Also added an API call to limit the maximum GC heap. This is useful
when you are using a ulimit on vmem since the OS simply kills the
process instead of making malloc calls return NULL.

The eval loop will now catch out of memory and stack overflow and
continue with the next command. This is annoying for mos-scripts but
is necessary to make OMPython not simply crash (we want to get the
error-message through getErrorString(), and hopefully there is enough
memory to get the string after collecting some garbage).