[NF] Relax variability error in functions
The binding expression of a parameter in a function is now allowed to be
a discrete expression (since we do not distinguish between inputs and
other bindings). This now gives a warning instead of an error (because
the Modelica specification does not seem to allow this, but other tools
and libraries use this pattern for no obvious reason).