[Cpp] Fix derivative vars in XmlPropertyReader, ticket:4773
Commit b528b72f1484c42d12f1be9b91cf16cca894ab7d
(Fatemeh Davoudi implementation of the symbolic model reduction algorithm)
introduced _derVars and _resVars along with a second constructor that
initialized the dimension of _derVars as _dimRHS. This second constructor
was not used though, leaving _dimRHS undefined and running a for loop with
undefined end during initialization.

This commit removes the second constructor and the uninitialized variable.
It uses consistent size info from SimVars instead.